Alterei o tipo do id

+++ Sistema/002-IMPLEMENTACAO/SISAP/src/main/java/br/edu/ifpb/monteiro/ads/sisap/entities/AtendimentoConverter.java
@@ -46,7 +46,7 @@ public class AtendimentoConverter implements Converter {
 		if (valor == null) {
 			return null;
 		}
-		Long matriculaAluno = ((Atendimento) valor).getId();
+		Integer matriculaAluno = ((Atendimento) valor).getId();
 
 		return (matriculaAluno != null) ? matriculaAluno.toString() : null;
 	}
